move primary_window_attributes into lib
